move performance test
